add navigation in the header and delete some un-wanted files
Delete all component test files as i will not be testing much on this site. if needed I will add them back later. routing is set up in app-routing and used in header component. I am using a pages array on the component then looping the path and name in the html.